+++ pkg/handlers/internalapi/moves.go
@@ -28,6 +28,75 @@ import (
 	"github.com/transcom/mymove/pkg/storage"
 )
 
+func payloadForPPMModel(storer storage.FileStorer, personallyProcuredMove models.PersonallyProcuredMove) (*internalmessages.PersonallyProcuredMovePayload, error) {
+
+	documentPayload, err := payloads.PayloadForDocumentModel(storer, personallyProcuredMove.AdvanceWorksheet)
+	var hasProGear *string
+	if personallyProcuredMove.HasProGear != nil {
+		hpg := string(*personallyProcuredMove.HasProGear)
+		hasProGear = &hpg
+	}
+	var hasProGearOverThousand *string
+	if personallyProcuredMove.HasProGearOverThousand != nil {
+		hpgot := string(*personallyProcuredMove.HasProGearOverThousand)
+		hasProGearOverThousand = &hpgot
+	}
+	if err != nil {
+		return nil, err
+	}
+	ppmPayload := internalmessages.PersonallyProcuredMovePayload{
+		ID:                            handlers.FmtUUID(personallyProcuredMove.ID),
+		MoveID:                        *handlers.FmtUUID(personallyProcuredMove.MoveID),
+		CreatedAt:                     handlers.FmtDateTime(personallyProcuredMove.CreatedAt),
+		UpdatedAt:                     handlers.FmtDateTime(personallyProcuredMove.UpdatedAt),
+		WeightEstimate:                handlers.FmtPoundPtr(personallyProcuredMove.WeightEstimate),
+		OriginalMoveDate:              handlers.FmtDatePtr(personallyProcuredMove.OriginalMoveDate),
+		ActualMoveDate:                handlers.FmtDatePtr(personallyProcuredMove.ActualMoveDate),
+		SubmitDate:                    handlers.FmtDateTimePtr(personallyProcuredMove.SubmitDate),
+		ApproveDate:                   handlers.FmtDateTimePtr(personallyProcuredMove.ApproveDate),
+		PickupPostalCode:              personallyProcuredMove.PickupPostalCode,
+		HasAdditionalPostalCode:       personallyProcuredMove.HasAdditionalPostalCode,
+		AdditionalPickupPostalCode:    personallyProcuredMove.AdditionalPickupPostalCode,
+		DestinationPostalCode:         personallyProcuredMove.DestinationPostalCode,
+		HasSit:                        personallyProcuredMove.HasSit,
+		DaysInStorage:                 personallyProcuredMove.DaysInStorage,
+		EstimatedStorageReimbursement: personallyProcuredMove.EstimatedStorageReimbursement,
+		Status:                        internalmessages.PPMStatus(personallyProcuredMove.Status),
+		HasRequestedAdvance:           &personallyProcuredMove.HasRequestedAdvance,
+		Advance:                       payloadForReimbursementModel(personallyProcuredMove.Advance),
+		AdvanceWorksheet:              documentPayload,
+		Mileage:                       personallyProcuredMove.Mileage,
+		TotalSitCost:                  handlers.FmtCost(personallyProcuredMove.TotalSITCost),
+		HasProGear:                    hasProGear,
+		HasProGearOverThousand:        hasProGearOverThousand,
+	}
+	if personallyProcuredMove.IncentiveEstimateMin != nil {
+		min := (*personallyProcuredMove.IncentiveEstimateMin).Int64()
+		ppmPayload.IncentiveEstimateMin = &min
+	}
+	if personallyProcuredMove.IncentiveEstimateMax != nil {
+		max := (*personallyProcuredMove.IncentiveEstimateMax).Int64()
+		ppmPayload.IncentiveEstimateMax = &max
+	}
+	if personallyProcuredMove.PlannedSITMax != nil {
+		max := (*personallyProcuredMove.PlannedSITMax).Int64()
+		ppmPayload.PlannedSitMax = &max
+	}
+	if personallyProcuredMove.SITMax != nil {
+		max := (*personallyProcuredMove.SITMax).Int64()
+		ppmPayload.SitMax = &max
+	}
+	if personallyProcuredMove.HasProGear != nil {
+		hasProGear := string(*personallyProcuredMove.HasProGear)
+		ppmPayload.HasProGear = &hasProGear
+	}
+	if personallyProcuredMove.HasProGearOverThousand != nil {
+		hasProGearOverThousand := string(*personallyProcuredMove.HasProGearOverThousand)
+		ppmPayload.HasProGearOverThousand = &hasProGearOverThousand
+	}
+	return &ppmPayload, nil
+}
+
 func payloadForMoveModel(storer storage.FileStorer, order models.Order, move models.Move) (*internalmessages.MovePayload, error) {
 
 	var ppmPayloads internalmessages.IndexPersonallyProcuredMovePayload
